GG Ol GG

+++ project-app/app/Http/Controllers/settingController.php
@@ -132,19 +132,27 @@ class settingController extends Controller
                      ->withInput(); // Fallback URL
                 }
                 break;
+            case 'customFields':
+                $department = department::findOrFail(Auth::user()->dept_id);
+                $deleterow = json_decode($department->custom_fields);
+
+                unset($deleterow[$id]);
+
+                $department->custom_fields = json_encode($deleterow);
+
+                $department->save();
+
+                break;
             default:
-                dd(["message" => "errors"]);
+            return redirect()->back()->withErrors('Failed to remove the item from the list.');
         }
 
 
-        if($deleteFrom !== null){
+        if($tab !== 'customFields' && $deleteFrom !== null){
             $deleteFrom->delete();
             return redirect()->back()->with('Setting deleted successfully.');
         }
-        else{
-            dd('Errors');
-            return redirect()->back()->withErrors('Setting deleted successfully.');
-        }
+
 
 
     }
@@ -201,7 +209,7 @@ class settingController extends Controller
 
 
                     if($column !== false ){
-                        return redirect()->back()->withErrors(['errors' =>'The Name Provided is existed']);
+                        return redirect()->back()->withErrors(['errors' =>'The Name already exists']);
                     }
